+++ Install/AltPresets.lua
@@ -1,5 +1,5 @@
 local _, ns = ...
 ns.InstallStrings = ns.InstallStrings or {}
 ns.InstallStrings.ALT_PRESETS = [==[
-!TUIALT2!TY5bDoIwEEW30iWoYQMGoWAA/cAFFDutk9S26QPDcvDDdVpjbPycM/fOmZUeQLCowrO1zszIwflXBWqOuNJEBCooh2akE8oHSp/hr1dJh3yX8aluRtJejfakVMz7fopCOLij/lzOsTrcUEtPk8g6nDMfzmliAch3f7SKhUWzYNyfuKr3l24cgGPoDYeOLSaGNj1JtkWxsW8=
+!TUIALT2!XY5RDoIwEESv0iOo4QYCBSOERC5Q0m3dWGnZthiOgx+e0xoTEv2ctzOzs/IclIgmPGtHdkYJ5F8FmDniyhNRaODYVj0fUD9Q+x9YaEJ52FBXVj27TFEQeJYLujVDVIrgjuOndvOV4Yqj9jx9cYTzxtsuKRGAfe8nZ0RYRhEs/UdbkBgaK+EsFhtDnbawfZbt3Bs=
 ]==]